refactor: move Crawler to main.py and use threads
